WIP: keycloak: same url in internal and external networks using root
Some client require configuration discovery of openid-connect. Sadly, the urls
are autogenerated by keycloak based on the host given by the web client.

This means that the url used by the web client and the client apps *must* be the
same on the host network and the internal network.

To do so:
- we add an alias on internal network so that keycloak can be accessed from the
same hostname
- we make keycloak run on port 80

To do the second point, we *must* be root (sadly... again)

TODO do we really need to be root? Would it be better to keep port 8080 to
everyone (since only admins need to access keycloak directly) ?